Make autocomplete.cc and dependencies compile on Posix. Review URL: http://codereview.chromium.org/28061
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@10277 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/autocomplete/history_contents_provider.h
@@ -37,7 +37,7 @@ class HistoryContentsProvider : public AutocompleteProvider {
 
   // The maximum match count we'll report. If the db_match_count is greater
   // than this, it will be clamped to this result.
-  static const int kMaxMatchCount = 50;
+  static const size_t kMaxMatchCount = 50;
 
  private:
   void QueryComplete(HistoryService::Handle handle,
